Moving formatter and helper base types onto root of addon namespace
