+++ addon/formatters/format-date.js
@@ -6,7 +6,7 @@
 import Ember from 'ember';
 import createFormatCache from 'intl-format-cache';
 
-import Formatter from './-base';
+import Formatter from '../formatter';
 
 const { assert, computed } = Ember;
