+++ addon/helpers/format-number.js
@@ -3,7 +3,7 @@
  * Copyrights licensed under the New BSD License. See the accompanying LICENSE file for terms.
  */
 
-import BaseHelper from './-format-base';
+import BaseHelper from '../helper';
 
 export default BaseHelper.extend({
   format(value, options) {
